# s2: Optimize for Intel LEA