Intel has a penalty for using LEA with displacement and less execution ports for LEA compared to ADD.

Use appropriate instructions when possible.

So far not too promising (AMD):
```
name                                      old speed      new speed      delta                      
TwainEncode1e1/default-32                 1.21GB/s ± 0%  1.21GB/s ± 0%      ~       (p=0.963 n=9+8)
TwainEncode1e1/better-32                  1.21GB/s ± 0%  1.21GB/s ± 0%      ~     (p=0.052 n=10+10)
TwainEncode1e1/best-32                    1.21GB/s ± 0%  1.21GB/s ± 0%      ~      (p=0.251 n=10+9)
TwainEncode1e2/default-32                 1.07GB/s ± 0%  1.05GB/s ± 0%    -1.68%  (p=0.000 n=10+10)
TwainEncode1e2/better-32                   371MB/s ± 0%   361MB/s ± 0%    -2.71%  (p=0.000 n=10+10)
TwainEncode1e2/best-32                    1.36MB/s ± 0%  1.35MB/s ± 0%    -0.37%   (p=0.013 n=8+10)
TwainEncode1e3/default-32                 1.15GB/s ± 0%  1.14GB/s ± 0%    -1.14%  (p=0.000 n=10+10)
TwainEncode1e3/better-32                   416MB/s ± 0%   413MB/s ± 0%    -0.74%  (p=0.000 n=10+10)
TwainEncode1e3/best-32                    7.97MB/s ± 0%  7.88MB/s ± 1%    -1.12%  (p=0.000 n=10+10)
TwainEncode1e4/default-32                 1.00GB/s ± 0%  0.99GB/s ± 0%    -1.12%  (p=0.000 n=10+10)
TwainEncode1e4/better-32                   418MB/s ± 0%   412MB/s ± 1%    -1.43%   (p=0.000 n=9+10)
TwainEncode1e4/best-32                    16.1MB/s ± 1%  15.8MB/s ± 1%    -1.63%  (p=0.000 n=10+10)
TwainEncode1e5/default-32                  487MB/s ± 0%   478MB/s ± 1%    -1.84%   (p=0.000 n=10+9)
TwainEncode1e5/better-32                   257MB/s ± 0%   251MB/s ± 0%    -2.21%   (p=0.000 n=9+10)
TwainEncode1e5/best-32                    19.9MB/s ± 0%  19.4MB/s ± 3%    -2.51%   (p=0.000 n=9+10)
TwainEncode1e6/default-32                 1.05GB/s ± 8%  1.06GB/s ± 1%      ~      (p=1.000 n=10+9)
TwainEncode1e6/better-32                   651MB/s ± 0%   636MB/s ± 1%    -2.19%    (p=0.000 n=8+9)
TwainEncode1e6/best-32                    52.7MB/s ± 3%  51.0MB/s ± 4%    -3.23%  (p=0.003 n=10+10)
RandomEncodeBlock1MB/default-32           36.8GB/s ± 0%  36.0GB/s ± 0%    -2.01%    (p=0.000 n=9+9)
RandomEncodeBlock1MB/better-32            15.9GB/s ± 0%  15.6GB/s ± 0%    -1.74%  (p=0.000 n=10+10)
RandomEncodeBlock1MB/best-32               797MB/s ± 0%   787MB/s ± 1%    -1.16%   (p=0.000 n=8+10)
RandomEncodeBetterBlock16MB/default-32    10.6GB/s ± 0%  10.2GB/s ± 4%    -3.83%  (p=0.000 n=10+10)
RandomEncodeBetterBlock16MB/better-32     5.96GB/s ± 1%  5.73GB/s ± 3%    -3.78%  (p=0.000 n=10+10)
RandomEncodeBetterBlock16MB/best-32        633MB/s ± 1%   605MB/s ± 3%    -4.39%   (p=0.000 n=8+10)
EncodeS2Block/0-html/block-32             27.9GB/s ± 1%  28.3GB/s ± 1%    +1.77%  (p=0.000 n=10+10)
EncodeS2Block/0-html/block-better-32      15.6GB/s ± 1%  15.6GB/s ± 0%    +0.51%  (p=0.019 n=10+10)
EncodeS2Block/0-html/block-best-32         375MB/s ± 1%   365MB/s ± 1%    -2.79%  (p=0.000 n=10+10)
EncodeS2Block/1-urls/block-32             15.5GB/s ± 1%  15.6GB/s ± 1%    +0.81%  (p=0.004 n=10+10)
EncodeS2Block/1-urls/block-better-32      8.08GB/s ± 1%  8.02GB/s ± 3%      ~     (p=0.912 n=10+10)
EncodeS2Block/1-urls/block-best-32         398MB/s ± 1%   394MB/s ± 3%      ~     (p=0.684 n=10+10)
EncodeS2Block/2-jpg/block-32               318GB/s ± 1%   322GB/s ± 1%    +1.16%  (p=0.000 n=10+10)
EncodeS2Block/2-jpg/block-better-32        134GB/s ± 1%   136GB/s ± 1%    +1.86%  (p=0.000 n=10+10)
EncodeS2Block/2-jpg/block-best-32          637MB/s ±12%   454MB/s ± 2%   -28.80%  (p=0.000 n=10+10)
EncodeS2Block/3-jpg_200b/block-32         25.1GB/s ± 1%  24.5GB/s ± 1%    -2.19%  (p=0.000 n=10+10)
EncodeS2Block/3-jpg_200b/block-better-32  10.2GB/s ± 1%  10.7GB/s ± 1%    +4.96%  (p=0.000 n=10+10)
EncodeS2Block/3-jpg_200b/block-best-32    1.95MB/s ±22%  0.77MB/s ± 6%   -60.40%  (p=0.000 n=10+10)
EncodeS2Block/4-pdf/block-32               136GB/s ± 1%   142GB/s ± 0%    +3.78%  (p=0.000 n=10+10)
EncodeS2Block/4-pdf/block-better-32       16.0GB/s ± 1%  16.5GB/s ± 1%    +3.01%   (p=0.000 n=9+10)
EncodeS2Block/4-pdf/block-best-32          277MB/s ± 1%   279MB/s ± 1%      ~     (p=0.075 n=10+10)
EncodeS2Block/5-html4/block-32            80.0GB/s ± 1%  84.8GB/s ± 1%    +6.01%  (p=0.000 n=10+10)
EncodeS2Block/5-html4/block-better-32     50.5GB/s ± 1%  52.3GB/s ± 0%    +3.43%  (p=0.000 n=10+10)
EncodeS2Block/5-html4/block-best-32       1.48GB/s ± 1%  1.44GB/s ± 1%    -2.29%   (p=0.000 n=10+8)
```